requirements.txt
absl-py==2.4.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
curl_cffi==0.13.0
dateutils==0.6.12
distro==1.9.0
dnspython==2.8.0
fastapi==0.128.0
flatbuffers==25.12.19
frozendict==2.4.7
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hmmlearn==0.3.3
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
llvmlite==0.44.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
multitasking==0.0.12
namex==0.1.0
networkx==3.6.1
nsetools==2.0.1
numba==0.61.2
numpy==2.2.6
nvidia-nccl-cu12==2.29.2
openai==2.16.0
opt_einsum==3.4.0
optree==0.18.0
optuna==4.7.0
packaging==26.0
pandas==3.0.0
pandas-ta==0.4.71b0
patsy==1.0.2
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotext==5.3.2
praw==7.8.1
prawcore==2.4.0
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow[and-cuda]
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.6.3
uvicorn==0.40.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.1
xgboost
yfinance==1.1.0
pytest==8.0.0
reportlab==4.1.0
